agent/requirements.txt
livekit-agents[codecs]~=1.0
livekit-plugins-silero~=1.0
python-dotenv~=1.1
